test(test-utils): consolidate test infrastructure and standardize spec locations
- Move generic mock factories from aiBeanImport/test-utils to src/test-utils
- Move generic bean helpers to src/test-utils/bean-test-helpers.ts
- Rename misnamed data factories: createMockBean -> createBean,
  createMockBeanParams -> createBeanParams, createMockSettings -> createSettings,
  createMockExamples -> createExamples
- Move 3 colocated specs into __tests__/ folders for consistency
- Keep AI-specific helpers in aiBeanImport/test-utils
- Update all affected imports
